requirements.txt
tensorflow==2.15.0
numpy
pillow
streamlit
scipy
matplotlib
